\name{TFIDF-package}
\alias{TFIDF-package}
\alias{TFIDF}
\docType{package}
\title{
\packageTitle{TFIDF}
}
\description{
\packageDescription{TFIDF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TFIDF}
\packageIndices{TFIDF}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{TFIDF}

Maintainer: \packageMaintainer{TFIDF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
